auth reply codes

+++ modules/auth.h
@@ -110,7 +110,7 @@ int auth_serv_token_get(struct iovec *tok, struct iovec *user, struct iovec *pas
 int auth_user_pass_verify(struct iovec *user, struct iovec *pass);
 int auth_user_token_new(auth_user_token_t *token, auth_payload_t *payload);
 int auth_user_token_set(char *userid, auth_user_token_t *token);
-int auth_user_token_use(struct iovec *token, struct iovec *pass);
+uint8_t auth_user_token_use(struct iovec *token, struct iovec *pass);
 int auth_user_token_valid(auth_user_token_t *token);
 int auth_decode_packet(lc_message_t *msg, auth_payload_t *payload);
 int auth_decode_packet_key(lc_message_t *msg, auth_payload_t *payload, unsigned char *sk);
